# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScopeGuard API specification - change order approval workflow for agencies
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ErrorResponseError = {
  code: string;
  message: string;
};

export interface ErrorResponse {
  error: ErrorResponseError;
}

export interface RegisterBody {
  organizationName: string;
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  ADMIN: "ADMIN",
  PROJECT_MANAGER: "PROJECT_MANAGER",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  organizationId: string;
  createdAt: string;
}

export interface Organization {
  id: string;
  name: string;
  slug: string;
  logoUrl?: string | null;
  primaryColor?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  organization: Organization;
}

export interface UpdateOrganizationBody {
  name?: string;
  logoUrl?: string | null;
  primaryColor?: string | null;
}

export interface ClientContact {
  id: string;
  name: string;
  email: string;
  company: string;
  organizationId: string;
  createdAt: string;
}

export interface CreateClientContactBody {
  name: string;
  email: string;
  company?: string;
}

export interface UpdateClientContactBody {
  name?: string;
  email?: string;
  company?: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  ARCHIVED: "ARCHIVED",
} as const;

export interface Project {
  id: string;
  name: string;
  description: string;
  status: ProjectStatus;
  /** Hourly rate stored in cents */
  hourlyRateCents: number;
  organizationId: string;
  clientContactId: string;
  clientName: string;
  clientEmail: string;
  clientCompany: string;
  /** Sum of all approved change request totals in cents */
  totalApprovedCents: number;
  changeRequestCount: number;
  pendingCount: number;
  approvedCount: number;
  createdAt: string;
  updatedAt: string;
}

export type ChangeRequestStatus =
  (typeof ChangeRequestStatus)[keyof typeof ChangeRequestStatus];

export const ChangeRequestStatus = {
  DRAFT: "DRAFT",
  SENT: "SENT",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISED: "REVISED",
} as const;

export interface ChangeRequest {
  id: string;
  projectId: string;
  projectName: string;
  clientName: string;
  clientEmail: string;
  title: string;
  description: string;
  estimatedHours: number;
  hourlyRateCents: number;
  totalCents: number;
  status: ChangeRequestStatus;
  deadline?: string | null;
  approvalToken: string;
  approvedAt?: string | null;
  clientComments: string;
  createdAt: string;
  updatedAt: string;
}

export type ProjectDetail = Project & {
  changeRequests: ChangeRequest[];
};

export interface CreateProjectBody {
  name: string;
  description?: string;
  clientContactId: string;
  /** @minimum 0 */
  hourlyRateCents: number;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  ARCHIVED: "ARCHIVED",
} as const;

export interface UpdateProjectBody {
  name?: string;
  description?: string;
  clientContactId?: string;
  /** @minimum 0 */
  hourlyRateCents?: number;
  status?: UpdateProjectBodyStatus;
}

export interface CreateChangeRequestBody {
  projectId: string;
  title: string;
  description: string;
  /** @minimum 0 */
  estimatedHours: number;
  deadline?: string | null;
}

export interface UpdateChangeRequestBody {
  title?: string;
  description?: string;
  /** @minimum 0 */
  estimatedHours?: number;
  deadline?: string | null;
}

export type PublicChangeRequestStatus =
  (typeof PublicChangeRequestStatus)[keyof typeof PublicChangeRequestStatus];

export const PublicChangeRequestStatus = {
  DRAFT: "DRAFT",
  SENT: "SENT",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISED: "REVISED",
} as const;

export interface PublicChangeRequest {
  projectName: string;
  organizationName: string;
  title: string;
  description: string;
  totalCents: number;
  totalFormatted: string;
  deadline?: string | null;
  logoUrl?: string | null;
  primaryColor?: string | null;
  status: PublicChangeRequestStatus;
}

export type SubmitApprovalBodyDecision =
  (typeof SubmitApprovalBodyDecision)[keyof typeof SubmitApprovalBodyDecision];

export const SubmitApprovalBodyDecision = {
  approve: "approve",
  reject: "reject",
  revise: "revise",
} as const;

export interface SubmitApprovalBody {
  decision: SubmitApprovalBodyDecision;
  comments?: string;
}

export interface ApprovalResponse {
  status: string;
  message: string;
}

export interface DashboardSummary {
  totalProjects: number;
  activeProjects: number;
  totalChangeRequests: number;
  pendingApproval: number;
  approvedThisMonth: number;
  totalApprovedCents: number;
  totalApprovedFormatted: string;
  draftCount: number;
  sentCount: number;
  approvedCount: number;
  rejectedCount: number;
  revisedCount: number;
}

export type ActivityItemAction =
  (typeof ActivityItemAction)[keyof typeof ActivityItemAction];

export const ActivityItemAction = {
  CREATED: "CREATED",
  SENT: "SENT",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISED: "REVISED",
  UPDATED: "UPDATED",
} as const;

export interface ActivityItem {
  id: string;
  changeRequestId: string;
  changeRequestTitle: string;
  projectName: string;
  clientName: string;
  action: ActivityItemAction;
  totalCents: number;
  totalFormatted: string;
  createdAt: string;
}

export type Logout200 = {
  message: string;
};

export type ListClientContactsParams = {
  search?: string;
};

export type ListProjectsParams = {
  search?: string;
  status?: ListProjectsStatus;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  ACTIVE: "ACTIVE",
  COMPLETED: "COMPLETED",
  ARCHIVED: "ARCHIVED",
} as const;

export type ListChangeRequestsParams = {
  projectId?: string;
  status?: ListChangeRequestsStatus;
  search?: string;
};

export type ListChangeRequestsStatus =
  (typeof ListChangeRequestsStatus)[keyof typeof ListChangeRequestsStatus];

export const ListChangeRequestsStatus = {
  DRAFT: "DRAFT",
  SENT: "SENT",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISED: "REVISED",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};