# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScopeGuard API specification - change order approval workflow for agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user and organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ORGANIZATIONS
...
  /approve/{token}:
    get:
      operationId: getApprovalRequest
      tags: [approval]
      summary: Get public change request details by approval token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public change request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicChangeRequest"
        "404":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: submitApproval
      tags: [approval]
      summary: Submit approval decision (approve, reject, or request revision)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitApprovalBody"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
        "400":
          description: Invalid decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]

    # AUTH SCHEMAS
    RegisterBody:
      type: object
      properties:
        organizationName:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [organizationName, name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        organization:
          $ref: "#/components/schemas/Organization"
      required: [user, organization]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, PROJECT_MANAGER]
        organizationId:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, organizationId, createdAt]

    # ORGANIZATION SCHEMAS
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt]
...
    PublicChangeRequest:
      type: object
      properties:
        projectName:
          type: string
        organizationName:
          type: string
        title:
          type: string
        description:
          type: string
        totalCents:
          type: integer
        totalFormatted:
          type: string
        deadline:
          type: string
          format: date
          nullable: true
        logoUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, REVISED]
      required: [projectName, organizationName, title, description, totalCents, totalFormatted, status]
    SubmitApprovalBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject, revise]
        comments:
          type: string
      required: [decision]
    ApprovalResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required: [status, message]

    # DASHBOARD SCHEMAS
...